requirements-dev.txt
aiocache==0.12.3
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
asyncer==0.0.8
backcall==0.2.0
backend==0.2.4.1
backports.tarfile==1.2.0
build==1.2.2.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
docutils==0.20.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
h11==0.14.0
id==1.5.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
keyring==25.5.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mypy==1.14.1
mypy-extensions==1.0.0
nh3==0.2.21
orjson==3.10.15
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyproject_hooks==1.2.0
readme_renderer==43.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
SecretStorage==3.3.3
snapapi==0.1.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.44.0
tomli==2.2.1
traitlets==5.14.3
twine==6.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
wcwidth==0.2.13
zipp==3.20.2
